1_Semana/Nivel_02/primeiro-projeto-node/src/modules/users/services/ResetPasswordService.ts
import UserToken from '@modules/users/infra/typeorm/entities/UserToken';
import IUserTokensRepository from '@modules/users/repositories/IUserTokensRepository';
import AppError from '@shared/errors/AppErro';
import { addHours, differenceInHours, isAfter } from 'date-fns';
import { inject, injectable } from 'tsyringe';
import IHashProvider from '../providers/HashProviders/models/IHashProvider';
import IUsersRepository from '../repositories/IUsersRepoisotiry';

interface IRequest {
  password: string;
  token: string;
}

@injectable()
class ResetPasswordService {
  constructor(
    @inject('UsersRepository')
    private userRepository: IUsersRepository,
    @inject('UserTokensRepository')
    private usersTokensRepository: IUserTokensRepository,
    @inject('HashProvider')
    private hashProvider: IHashProvider,
  ) {}

  public async execute({ password, token }: IRequest): Promise<void> {
    const userToken = await this.usersTokensRepository.findByToken(token);
    if (!userToken) {
      throw new AppError('User Token does not exists');
    }
    const user = await this.userRepository.findById(userToken.user_id);
    if (!user) {
      throw new AppError('User does not exists');
    }

    const tokenCreatedAt = userToken.created_at;
    if (isAfter(Date.now(), addHours(tokenCreatedAt, 2))) {
      throw new AppError('Token expired.');
    }
    user.password = await this.hashProvider.generateHash(password);
  }
}

export default ResetPasswordService;
